requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3
sqlalchemy==1.4.46
apache-airflow-providers-postgres==5.6.0
python-dotenv==0.21.0